refactor: move import symbols to feature
