+++ packages/core/src/features/css-type.ts
@@ -1,7 +1,8 @@
 import { createFeature, FeatureContext } from './feature';
-import type { StylableDirectives, ImportSymbol } from './types';
+import type { StylableDirectives } from './types';
 import { generalDiagnostics } from './diagnostics';
 import * as STSymbol from './st-symbol';
+import type { ImportSymbol } from './st-import';
 import * as CSSClass from './css-class';
 import { plugableRecord } from '../helpers/plugable-record';
 import type { StylableMeta } from '../stylable-meta';
