+++ packages/core/src/features/st-symbol.ts
@@ -1,5 +1,6 @@
 import { FeatureContext, createFeature } from './feature';
-import type { ImportSymbol, VarSymbol, CSSVarSymbol, KeyframesSymbol } from './types';
+import type { VarSymbol, CSSVarSymbol, KeyframesSymbol } from './types';
+import type { ImportSymbol } from './st-import';
 import type { ClassSymbol } from './css-class';
 import type { ElementSymbol } from './css-type';
 import { plugableRecord } from '../helpers/plugable-record';
